Fixing window size

+++ src/lcd_gui_qt/lcd_draw_properties.h
@@ -9,18 +9,18 @@ namespace lcd
 
 	struct lcd_draw_properties
 	{
-		QPainter* painter;
-		i_character_data* char_data;
+		QPainter*									painter;
+		i_character_data*							char_data;
 		std::array<std::array<char, 8>, 255> const* font;
-		size_t lcd_width;
-		size_t lcd_height;
-		size_t rows;
-		size_t columns;
-		size_t char_width;
-		size_t char_height;
-		size_t char_hspacing;
-		size_t char_vspacing;
-		size_t pixel_size;
-		size_t pixel_spacing;
+		size_t										lcd_width;
+		size_t										lcd_height;
+		size_t										rows;
+		size_t										columns;
+		float										char_width;
+		float										char_height;
+		float										char_hspacing;
+		float										char_vspacing;
+		float										pixel_size;
+		float										pixel_spacing;
 	};
-}
+} // namespace lcd
